Always use randomized prompt
